+++ src/npctalk.cpp
@@ -3847,8 +3847,8 @@ std::string give_item_to( npc &p, bool allow_use, bool allow_carry )
     }
 
     bool taken = false;
-    long our_ammo = p.ammo_count_for( p.weapon );
-    long new_ammo = p.ammo_count_for( given );
+    int our_ammo = p.ammo_count_for( p.weapon );
+    int new_ammo = p.ammo_count_for( given );
     const double new_weapon_value = p.weapon_value( given, new_ammo );
     const double cur_weapon_value = p.weapon_value( p.weapon, our_ammo );
     if( allow_use ) {
